refactor: consolidate automation model files
- merge automation_run.rs / automation_run_event.rs / automation_trigger.rs into automation.rs
- upgrade repository/automation.rs section comments to // ── name ── style
