+++ backend/src/model/mod.rs
@@ -1,7 +1,4 @@
 mod automation;
-mod automation_run;
-mod automation_run_event;
-mod automation_trigger;
 mod dirent;
 mod document;
 mod project;
@@ -9,9 +6,6 @@ mod session;
 mod user;
 
 pub use automation::*;
-pub use automation_run::*;
-pub use automation_run_event::*;
-pub use automation_trigger::*;
 pub use dirent::*;
 pub use document::*;
 pub use project::*;
